+++ client/src/App.js
@@ -19,7 +19,7 @@ import Lobby from './components/lobby/lobby';
 import Confirm from './components/confirm/confirm';
 import NotFound from './components/pagenotfound/page-not-found';
 import Contrib from './components/contrib/contrib.js';
-import { ChangePassword, ResetPassword } from './components/password/password';
+import { UpdatePassword, ResetPassword } from './components/password/password';
 import PrivateRoute from './components/private/private';
 
 class App extends Component {
@@ -66,7 +66,7 @@ class App extends Component {
             <Route exact path = '/contributors' component = { Contrib } />
             <Route path = '/@/:profileId' component = { Profile } />
             <Route path = '/password/reset' component = { ResetPassword } />
-            <PrivateRoute path = '/password/update' component = { ChangePassword } /> 
+            <PrivateRoute path = '/password/update' component = { UpdatePassword } /> 
             <PrivateRoute path = '/play' component = { Lobby } />
             <Route component = { NotFound } />
