+++ client/src/components/password/password.css
@@ -1,25 +1,26 @@
-.pswd-title{
-    color: #8299fd;
-    text-align: center;
+.pswd-title {
+  color: #8299fd;
+  text-align: center;
 }
 
 .mybtn {
-    background-color: #8299fd;
-    color: #ffffff;
+  background-color: #8299fd;
+  color: #ffffff;
 }
 
-.mybtn:hover, .mybtn:focus {
-    background-color: #8198fd99;
+.mybtn:hover,
+.mybtn:focus {
+  background-color: #8198fd99;
 }
 
 label {
-    display: block;
-    margin-top: 10px;
+  display: block;
+  margin-top: 10px;
 }
 
 .password-card {
-    max-width: 350px !important;
-    padding: 40px 40px;
-    margin: 0 auto 25px;
-    margin-top: 50px;
+  max-width: 350px !important;
+  padding: 40px 40px;
+  margin: 0 auto 25px;
+  margin-top: 50px;
 }
